requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
sentence-transformers
openai
requests
pydantic
rich
pytest
